Move summary title to own component
