+++ context/app/static/js/components/detailPage/summary/SummaryTitle/index.js
@@ -1,0 +1,3 @@
+import SummaryTitle from './SummaryTitle';
+
+export default SummaryTitle;
